+++ tests/datasource/fluent/data_asset/test_file_path_asset.py
@@ -13,7 +13,7 @@ from great_expectations.core.partitioners import (
     PartitionerYearly,
 )
 from great_expectations.datasource.fluent.data_asset.data_connector import FilePathDataConnector
-from great_expectations.datasource.fluent.file_path_asset_mixins import (
+from great_expectations.datasource.fluent.file_path_asset_base import (
     AmbiguousPathError,
     PathNotFoundError,
     RegexMissingRequiredGroupsError,
